[core] Introduce action factory (#1528)
