+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/CompactAction.java
@@ -27,29 +27,18 @@ import org.apache.paimon.table.AppendOnlyFileStoreTable;
 import org.apache.paimon.table.FileStoreTable;
 
 import org.apache.flink.api.common.RuntimeExecutionMode;
-import org.apache.flink.api.java.tuple.Tuple3;
-import org.apache.flink.api.java.utils.MultipleParameterTool;
 import org.apache.flink.configuration.ExecutionOptions;
 import org.apache.flink.configuration.ReadableConfig;
 import org.apache.flink.streaming.api.datastream.DataStreamSource;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.table.data.RowData;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-
-import static org.apache.paimon.flink.action.Action.getPartitions;
-import static org.apache.paimon.flink.action.Action.getTablePath;
-import static org.apache.paimon.flink.action.Action.optionalConfigMap;
 
 /** Table compact action for Flink. */
 public class CompactAction extends TableActionBase {
-
-    private static final Logger LOG = LoggerFactory.getLogger(CompactAction.class);
 
     private List<Map<String, String>> partitions;
 
@@ -122,71 +111,6 @@ public class CompactAction extends TableActionBase {
         unawareBucketCompactionTopoBuilder.withPartitions(partitions);
         unawareBucketCompactionTopoBuilder.withContinuousMode(isStreaming);
         unawareBucketCompactionTopoBuilder.build();
-    }
-
-    // ------------------------------------------------------------------------
-    //  Flink run methods
-    // ------------------------------------------------------------------------
-
-    public static Optional<Action> create(String[] args) {
-        LOG.info("Compact job args: {}", String.join(" ", args));
-
-        MultipleParameterTool params = MultipleParameterTool.fromArgs(args);
-
-        if (params.has("help")) {
-            printHelp();
-            return Optional.empty();
-        }
-
-        Tuple3<String, String, String> tablePath = getTablePath(params);
-
-        Map<String, String> catalogConfig = optionalConfigMap(params, "catalog-conf");
-
-        CompactAction action =
-                new CompactAction(tablePath.f0, tablePath.f1, tablePath.f2, catalogConfig);
-
-        if (params.has("partition")) {
-            List<Map<String, String>> partitions = getPartitions(params);
-            action.withPartitions(partitions);
-        }
-
-        return Optional.of(action);
-    }
-
-    private static void printHelp() {
-        System.out.println(
-                "Action \"compact\" runs a dedicated job for compacting specified table.");
-        System.out.println();
-
-        System.out.println("Syntax:");
-        System.out.println(
-                "  compact --warehouse <warehouse-path> --database <database-name> "
-                        + "--table <table-name> [--partition <partition-name>]");
-        System.out.println(
-                "  compact --warehouse s3://path/to/warehouse --database <database-name> "
-                        + "--table <table-name> [--catalog-conf <paimon-catalog-conf> [--catalog-conf <paimon-catalog-conf> ...]]");
-        System.out.println("  compact --path <table-path> [--partition <partition-name>]");
-        System.out.println();
-
-        System.out.println("Partition name syntax:");
-        System.out.println("  key1=value1,key2=value2,...");
-        System.out.println();
-
-        System.out.println("Examples:");
-        System.out.println(
-                "  compact --warehouse hdfs:///path/to/warehouse --database test_db --table test_table");
-        System.out.println(
-                "  compact --path hdfs:///path/to/warehouse/test_db.db/test_table --partition dt=20221126,hh=08");
-        System.out.println(
-                "  compact --warehouse hdfs:///path/to/warehouse --database test_db --table test_table "
-                        + "--partition dt=20221126,hh=08 --partition dt=20221127,hh=09");
-        System.out.println(
-                "  compact --warehouse s3:///path/to/warehouse "
-                        + "--database test_db "
-                        + "--table test_table "
-                        + "--catalog-conf s3.endpoint=https://****.com "
-                        + "--catalog-conf s3.access-key=***** "
-                        + "--catalog-conf s3.secret-key=***** ");
     }
 
     @Override
