+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/DeleteAction.java
@@ -18,8 +18,6 @@
 
 package org.apache.paimon.flink.action;
 
-import org.apache.flink.api.java.tuple.Tuple3;
-import org.apache.flink.api.java.utils.MultipleParameterTool;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.table.api.Table;
 import org.apache.flink.table.data.GenericRowData;
@@ -32,11 +30,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
-
-import static org.apache.paimon.flink.action.Action.getTablePath;
-import static org.apache.paimon.flink.action.Action.optionalConfigMap;
 
 /** Delete from table action for Flink. */
 public class DeleteAction extends TableActionBase {
@@ -54,54 +48,6 @@ public class DeleteAction extends TableActionBase {
         super(warehouse, databaseName, tableName, catalogConfig);
         changeIgnoreMergeEngine();
         this.filter = filter;
-    }
-
-    public static Optional<Action> create(String[] args) {
-        LOG.info("Delete job args: {}", String.join(" ", args));
-
-        MultipleParameterTool params = MultipleParameterTool.fromArgs(args);
-
-        if (params.has("help")) {
-            printHelp();
-            return Optional.empty();
-        }
-
-        Tuple3<String, String, String> tablePath = getTablePath(params);
-
-        String filter = params.get("where");
-        if (filter == null) {
-            throw new IllegalArgumentException(
-                    "Please specify deletion filter. If you want to delete all records, please use overwrite (see doc).");
-        }
-
-        Map<String, String> catalogConfig = optionalConfigMap(params, "catalog-conf");
-
-        DeleteAction action =
-                new DeleteAction(tablePath.f0, tablePath.f1, tablePath.f2, filter, catalogConfig);
-
-        return Optional.of(action);
-    }
-
-    private static void printHelp() {
-        System.out.println("Action \"delete\" deletes data from a table.");
-        System.out.println();
-
-        System.out.println("Syntax:");
-        System.out.println(
-                "  delete --warehouse <warehouse-path> --database <database-name> "
-                        + "--table <table-name> --where <filter_spec>");
-        System.out.println("  delete --path <table-path> --where <filter_spec>");
-        System.out.println();
-
-        System.out.println(
-                "The '--where <filter_spec>' part is equal to the 'WHERE' clause in SQL DELETE statement. If you want to delete all records, please use overwrite (see doc).");
-        System.out.println();
-
-        System.out.println("Examples:");
-        System.out.println(
-                "  delete --path hdfs:///path/to/warehouse/test_db.db/test_table --where id > (SELECT count(*) FROM employee)");
-        System.out.println(
-                "  It's equal to 'DELETE FROM test_table WHERE id > (SELECT count(*) FROM employee)");
     }
 
     @Override
