+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/cdc/kafka/KafkaSyncDatabaseAction.java
@@ -33,7 +33,6 @@ import org.apache.paimon.table.FileStoreTable;
 import org.apache.paimon.utils.Preconditions;
 
 import org.apache.flink.api.common.eventtime.WatermarkStrategy;
-import org.apache.flink.api.java.utils.MultipleParameterTool;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.connector.kafka.source.KafkaSource;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
@@ -49,13 +48,10 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.function.Supplier;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import static org.apache.paimon.flink.action.Action.checkRequiredArgument;
-import static org.apache.paimon.flink.action.Action.optionalConfigMap;
 import static org.apache.paimon.utils.Preconditions.checkArgument;
 
 /**
@@ -332,132 +328,6 @@ public class KafkaSyncDatabaseAction extends ActionBase {
     // ------------------------------------------------------------------------
     //  Flink run methods
     // ------------------------------------------------------------------------
-
-    public static Optional<Action> create(String[] args) {
-        MultipleParameterTool params = MultipleParameterTool.fromArgs(args);
-
-        if (params.has("help")) {
-            printHelp();
-            return Optional.empty();
-        }
-
-        checkRequiredArgument(params, "warehouse");
-        checkRequiredArgument(params, "database");
-        checkRequiredArgument(params, "kafka-conf");
-
-        int schemaInitMaxRead = 1000;
-        if (params.has("schema-init-max-read")) {
-            schemaInitMaxRead = Integer.parseInt(params.get("schema-init-max-read"));
-        }
-
-        String warehouse = params.get("warehouse");
-        String database = params.get("database");
-        boolean ignoreIncompatible = Boolean.parseBoolean(params.get("ignore-incompatible"));
-        String tablePrefix = params.get("table-prefix");
-        String tableSuffix = params.get("table-suffix");
-        String includingTables = params.get("including-tables");
-        String excludingTables = params.get("excluding-tables");
-
-        Map<String, String> kafkaConfigOption = optionalConfigMap(params, "kafka-conf");
-        Map<String, String> catalogConfigOption = optionalConfigMap(params, "catalog-conf");
-        Map<String, String> tableConfigOption = optionalConfigMap(params, "table-conf");
-        return Optional.of(
-                new KafkaSyncDatabaseAction(
-                        kafkaConfigOption,
-                        warehouse,
-                        database,
-                        schemaInitMaxRead,
-                        ignoreIncompatible,
-                        tablePrefix,
-                        tableSuffix,
-                        includingTables,
-                        excludingTables,
-                        catalogConfigOption,
-                        tableConfigOption));
-    }
-
-    private static void printHelp() {
-        System.out.println(
-                "Action \"kafka-sync-database\" creates a streaming job "
-                        + "with a Flink Kafka source and multiple Paimon table sinks "
-                        + "to synchronize multiple tables into one Paimon database.\n"
-                        + "Only tables with primary keys will be considered. ");
-        System.out.println();
-
-        System.out.println("Syntax:");
-        System.out.println(
-                "  kafka-sync-database --warehouse <warehouse-path> --database <database-name> "
-                        + "[--schema-init-max-read <schema-init-max-read>] "
-                        + "[--ignore-incompatible <true/false>] "
-                        + "[--table-prefix <paimon-table-prefix>] "
-                        + "[--table-suffix <paimon-table-suffix>] "
-                        + "[--including-tables <table-name|name-regular-expr>] "
-                        + "[--excluding-tables <table-name|name-regular-expr>] "
-                        + "[--kafka-conf <kafka-source-conf> [--kafka-conf <kafka-source-conf> ...]] "
-                        + "[--catalog-conf <paimon-catalog-conf> [--catalog-conf <paimon-catalog-conf> ...]] "
-                        + "[--table-conf <paimon-table-sink-conf> [--table-conf <paimon-table-sink-conf> ...]]");
-        System.out.println();
-
-        System.out.println(
-                "--schema-init-max-read is default 1000, if your tables are all from a topic, you can set this parameter to initialize the number of tables to be synchronized.");
-        System.out.println();
-
-        System.out.println(
-                "--ignore-incompatible is default false, in this case, if Topic's table name exists in Paimon "
-                        + "and their schema is incompatible, an exception will be thrown. "
-                        + "You can specify it to true explicitly to ignore the incompatible tables and exception.");
-        System.out.println();
-
-        System.out.println(
-                "--table-prefix is the prefix of all Paimon tables to be synchronized. For example, if you want all "
-                        + "synchronized tables to have \"ods_\" as prefix, you can specify `--table-prefix ods_`.");
-        System.out.println("The usage of --table-suffix is same as `--table-prefix`");
-        System.out.println();
-
-        System.out.println(
-                "--including-tables is used to specify which source tables are to be synchronized. "
-                        + "You must use '|' to separate multiple tables. Regular expression is supported.");
-        System.out.println(
-                "--excluding-tables is used to specify which source tables are not to be synchronized. "
-                        + "The usage is same as --including-tables.");
-        System.out.println(
-                "--excluding-tables has higher priority than --including-tables if you specified both.");
-        System.out.println();
-
-        System.out.println("kafka source conf syntax:");
-        System.out.println("  key=value");
-        System.out.println(
-                "'topic', 'properties.bootstrap.servers', 'properties.group.id'"
-                        + "are required configurations, others are optional.");
-        System.out.println(
-                "For a complete list of supported configurations, "
-                        + "see https://nightlies.apache.org/flink/flink-docs-release-1.16/zh/docs/connectors/table/kafka/");
-        System.out.println();
-        System.out.println();
-
-        System.out.println("Paimon catalog and table sink conf syntax:");
-        System.out.println("  key=value");
-        System.out.println("All Paimon sink table will be applied the same set of configurations.");
-        System.out.println(
-                "For a complete list of supported configurations, "
-                        + "see https://paimon.apache.org/docs/master/maintenance/configurations/");
-        System.out.println();
-
-        System.out.println("Examples:");
-        System.out.println(
-                "  kafka-sync-database \\\n"
-                        + "    --warehouse hdfs:///path/to/warehouse \\\n"
-                        + "    --database test_db \\\n"
-                        + "    --kafka-conf properties.bootstrap.servers=127.0.0.1:9020 \\\n"
-                        + "    --kafka-conf topic=order,logistic,user \\\n"
-                        + "    --kafka-conf properties.group.id=123456 \\\n"
-                        + "    --kafka-conf value.format=canal-json \\\n"
-                        + "    --catalog-conf metastore=hive \\\n"
-                        + "    --catalog-conf uri=thrift://hive-metastore:9083 \\\n"
-                        + "    --table-conf bucket=4 \\\n"
-                        + "    --table-conf changelog-producer=input \\\n"
-                        + "    --table-conf sink.parallelism=4");
-    }
 
     @Override
     public void run() throws Exception {
