+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/cdc/mysql/MySqlSyncDatabaseAction.java
@@ -34,7 +34,6 @@ import org.apache.paimon.utils.Preconditions;
 import com.ververica.cdc.connectors.mysql.source.MySqlSource;
 import com.ververica.cdc.connectors.mysql.source.config.MySqlSourceOptions;
 import org.apache.flink.api.common.eventtime.WatermarkStrategy;
-import org.apache.flink.api.java.utils.MultipleParameterTool;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.slf4j.Logger;
@@ -51,13 +50,10 @@ import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.function.Supplier;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import static org.apache.paimon.flink.action.Action.checkRequiredArgument;
-import static org.apache.paimon.flink.action.Action.optionalConfigMap;
 import static org.apache.paimon.flink.action.cdc.mysql.MySqlDatabaseSyncMode.SEPARATE;
 import static org.apache.paimon.flink.action.cdc.mysql.MySqlDatabaseSyncMode.UNIFIED;
 import static org.apache.paimon.utils.Preconditions.checkArgument;
@@ -346,132 +342,6 @@ public class MySqlSyncDatabaseAction extends ActionBase {
     // ------------------------------------------------------------------------
     //  Flink run methods
     // ------------------------------------------------------------------------
-
-    public static Optional<Action> create(String[] args) {
-        MultipleParameterTool params = MultipleParameterTool.fromArgs(args);
-
-        if (params.has("help")) {
-            printHelp();
-            return Optional.empty();
-        }
-
-        checkRequiredArgument(params, "warehouse");
-        checkRequiredArgument(params, "database");
-        checkRequiredArgument(params, "mysql-conf");
-
-        String warehouse = params.get("warehouse");
-        String database = params.get("database");
-        boolean ignoreIncompatible = Boolean.parseBoolean(params.get("ignore-incompatible"));
-        String tablePrefix = params.get("table-prefix");
-        String tableSuffix = params.get("table-suffix");
-        String includingTables = params.get("including-tables");
-        String excludingTables = params.get("excluding-tables");
-        String mode = params.get("mode");
-        MySqlDatabaseSyncMode syncMode;
-        if ("unified".equalsIgnoreCase(mode)) {
-            syncMode = UNIFIED;
-        } else {
-            syncMode = SEPARATE;
-        }
-
-        Map<String, String> mySqlConfig = optionalConfigMap(params, "mysql-conf");
-        Map<String, String> catalogConfig = optionalConfigMap(params, "catalog-conf");
-        Map<String, String> tableConfig = optionalConfigMap(params, "table-conf");
-        return Optional.of(
-                new MySqlSyncDatabaseAction(
-                        mySqlConfig,
-                        warehouse,
-                        database,
-                        ignoreIncompatible,
-                        tablePrefix,
-                        tableSuffix,
-                        includingTables,
-                        excludingTables,
-                        catalogConfig,
-                        tableConfig,
-                        syncMode));
-    }
-
-    private static void printHelp() {
-        System.out.println(
-                "Action \"mysql-sync-database\" creates a streaming job "
-                        + "with a Flink MySQL CDC source and multiple Paimon table sinks "
-                        + "to synchronize a whole MySQL database into one Paimon database.\n"
-                        + "Only MySQL tables with primary keys will be considered. "
-                        + "Newly created MySQL tables after the job starts will not be included.");
-        System.out.println();
-
-        System.out.println("Syntax:");
-        System.out.println(
-                "  mysql-sync-database --warehouse <warehouse-path> --database <database-name> "
-                        + "[--ignore-incompatible <true/false>] "
-                        + "[--table-prefix <paimon-table-prefix>] "
-                        + "[--table-suffix <paimon-table-suffix>] "
-                        + "[--including-tables <mysql-table-name|name-regular-expr>] "
-                        + "[--excluding-tables <mysql-table-name|name-regular-expr>] "
-                        + "[--mysql-conf <mysql-cdc-source-conf> [--mysql-conf <mysql-cdc-source-conf> ...]] "
-                        + "[--catalog-conf <paimon-catalog-conf> [--catalog-conf <paimon-catalog-conf> ...]] "
-                        + "[--table-conf <paimon-table-sink-conf> [--table-conf <paimon-table-sink-conf> ...]]");
-        System.out.println();
-
-        System.out.println(
-                "--ignore-incompatible is default false, in this case, if MySQL table name exists in Paimon "
-                        + "and their schema is incompatible, an exception will be thrown. "
-                        + "You can specify it to true explicitly to ignore the incompatible tables and exception.");
-        System.out.println();
-
-        System.out.println(
-                "--table-prefix is the prefix of all Paimon tables to be synchronized. For example, if you want all "
-                        + "synchronized tables to have \"ods_\" as prefix, you can specify `--table-prefix ods_`.");
-        System.out.println("The usage of --table-suffix is same as `--table-prefix`");
-        System.out.println();
-
-        System.out.println(
-                "--including-tables is used to specify which source tables are to be synchronized. "
-                        + "You must use '|' to separate multiple tables. Regular expression is supported.");
-        System.out.println(
-                "--excluding-tables is used to specify which source tables are not to be synchronized. "
-                        + "The usage is same as --including-tables.");
-        System.out.println(
-                "--excluding-tables has higher priority than --including-tables if you specified both.");
-        System.out.println();
-
-        System.out.println("MySQL CDC source conf syntax:");
-        System.out.println("  key=value");
-        System.out.println(
-                "'hostname', 'username', 'password' and 'database-name' "
-                        + "are required configurations, others are optional. "
-                        + "Note that 'database-name' should be the exact name "
-                        + "of the MySQL database you want to synchronize. "
-                        + "It can't be a regular expression.");
-        System.out.println(
-                "For a complete list of supported configurations, "
-                        + "see https://ververica.github.io/flink-cdc-connectors/master/content/connectors/mysql-cdc.html#connector-options");
-        System.out.println();
-
-        System.out.println("Paimon catalog and table sink conf syntax:");
-        System.out.println("  key=value");
-        System.out.println("All Paimon sink table will be applied the same set of configurations.");
-        System.out.println(
-                "For a complete list of supported configurations, "
-                        + "see https://paimon.apache.org/docs/master/maintenance/configurations/");
-        System.out.println();
-
-        System.out.println("Examples:");
-        System.out.println(
-                "  mysql-sync-database \\\n"
-                        + "    --warehouse hdfs:///path/to/warehouse \\\n"
-                        + "    --database test_db \\\n"
-                        + "    --mysql-conf hostname=127.0.0.1 \\\n"
-                        + "    --mysql-conf username=root \\\n"
-                        + "    --mysql-conf password=123456 \\\n"
-                        + "    --mysql-conf database-name=source_db \\\n"
-                        + "    --catalog-conf metastore=hive \\\n"
-                        + "    --catalog-conf uri=thrift://hive-metastore:9083 \\\n"
-                        + "    --table-conf bucket=4 \\\n"
-                        + "    --table-conf changelog-producer=input \\\n"
-                        + "    --table-conf sink.parallelism=4");
-    }
 
     @Override
     public void run() throws Exception {
