+++ paimon-flink/paimon-flink-common/src/main/resources/META-INF/services/org.apache.paimon.factories.Factory
@@ -14,3 +14,14 @@
 # limitations under the License.
 
 org.apache.paimon.flink.kafka.KafkaLogStoreFactory
+org.apache.paimon.flink.action.CompactActionFactory
+org.apache.paimon.flink.action.DropPartitionActionFactory
+org.apache.paimon.flink.action.DeleteActionFactory
+org.apache.paimon.flink.action.MergeIntoActionFactory
+org.apache.paimon.flink.action.RollbackToActionFactory
+org.apache.paimon.flink.action.CreateTagActionFactory
+org.apache.paimon.flink.action.DeleteTagActionFactory
+org.apache.paimon.flink.action.cdc.mysql.MySqlSyncTableActionFactory
+org.apache.paimon.flink.action.cdc.mysql.MySqlSyncDatabaseActionFactory
+org.apache.paimon.flink.action.cdc.kafka.KafkaSyncTableActionFactory
+org.apache.paimon.flink.action.cdc.kafka.KafkaSyncDatabaseActionFactory
